#41 feat: create, update, delete 권한 추가

+++ src/main/java/com/example/mangoplace/domain/review/controller/ReviewController.java
@@ -1,17 +1,24 @@
 package com.example.mangoplace.domain.review.controller;
 
+import com.example.mangoplace.domain.auth.security.AuthTokenFilter;
+import com.example.mangoplace.domain.auth.security.JwtUtils;
 import com.example.mangoplace.domain.review.dto.request.CreateReviewRequest;
 import com.example.mangoplace.domain.review.dto.request.UpdateReviewRequest;
 import com.example.mangoplace.domain.review.dto.response.CreateReviewResponse;
 import com.example.mangoplace.domain.review.dto.response.DeleteReviewResponse;
 import com.example.mangoplace.domain.review.dto.response.ReviewResponse;
 import com.example.mangoplace.domain.review.dto.response.UpdateReviewResponse;
+import com.example.mangoplace.domain.review.entity.Review;
 import com.example.mangoplace.domain.review.service.ReviewService;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.context.request.RequestContextHolder;
+import org.springframework.web.context.request.ServletRequestAttributes;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
@@ -23,6 +30,8 @@ import java.util.List;
 public class ReviewController {
 
     private final ReviewService reviewService;
+    private final JwtUtils jwtUtils;
+    private final AuthTokenFilter authTokenFilter;
     private final ObjectMapper objectMapper;
 
     @GetMapping("/{restaurantId}")
@@ -34,6 +43,7 @@ public class ReviewController {
     }
 
 
+    @PreAuthorize("hasRole('USER')")
     @PostMapping("/{restaurantId}")
     public ResponseEntity<CreateReviewResponse> createReview(
             @PathVariable String restaurantId,
@@ -60,24 +70,50 @@ public class ReviewController {
 
 
     @PutMapping("/{reviewId}")
+    @PreAuthorize("@reviewControllerSecurity.checkReviewOwnership(#reviewId)")
     public ResponseEntity<UpdateReviewResponse> updateReview(
             @PathVariable Long reviewId,
             @ModelAttribute UpdateReviewRequest updateReviewRequest,
             @RequestParam(value = "images", required = false) List<MultipartFile> images) {
 
         try {
-            UpdateReviewResponse response = reviewService.updateReview(reviewId, updateReviewRequest, images);
-            return ResponseEntity.status(HttpStatus.OK).body(response);
+            String jwtToken = getJwt();
+            String currentUsername = jwtUtils.getUserNameFromJwtToken(jwtToken);
+
+            Review review = reviewService.getReviewById(reviewId);
+
+            if (review != null && currentUsername.equals(review.getUser().getUsername())) {
+                UpdateReviewResponse response = reviewService.updateReview(reviewId, updateReviewRequest, images);
+                return ResponseEntity.status(HttpStatus.OK).body(response);
+            } else {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
         } catch (IOException e) {
             // 이미지 업로드 중 오류 발생 시 처리
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
         }
     }
 
-
     @DeleteMapping("/{reviewId}")
+    @PreAuthorize("@reviewControllerSecurity.checkReviewOwnership(#reviewId)")
     public ResponseEntity<DeleteReviewResponse> deleteReview(@PathVariable Long reviewId) {
-        DeleteReviewResponse response = reviewService.deleteReview(reviewId);
-        return ResponseEntity.ok(response);
+
+        String jwtToken = getJwt();
+        String currentUsername = jwtUtils.getUserNameFromJwtToken(jwtToken);
+
+        Review review = reviewService.getReviewById(reviewId);
+
+        if (review != null && currentUsername.equals(review.getUser().getUsername())) {
+            DeleteReviewResponse response = reviewService.deleteReview(reviewId);
+            return ResponseEntity.ok(response);
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+    }
+
+    private String getJwt(){
+        HttpServletRequest request = ((ServletRequestAttributes)
+                RequestContextHolder.getRequestAttributes()).getRequest();
+        return authTokenFilter.parseJwt(request);
     }
 }
